fix the logout functionality and separate the controllers for user
